misc: l3g4200d: Keep track of the sensor config
Read and store the initial configuration of the sensor and use the
local copy to update the sensor. This allows to avoid unnecessary
reads and ensures that the sensor is always configured the way we
expect. Also, update the local variables only after successul writes
and not before that, otherwise we might skip some register updates
thinking that the sensor is already configured. This can currently
happen when updating REG1 to change the ODR.

Change-Id: I9951eff6f4ed2d50a87b35d28a82df61275d5499